skims\feat(back): #6198.3 download root from presigned url
- download and extract repos

+++ skims/skims/batch/repositories.py
@@ -1,3 +1,7 @@
+import aiohttp
+from aiohttp.client_exceptions import (
+    ClientResponseError,
+)
 import asyncio
 from ctx import (
     NAMESPACES_FOLDER,
@@ -17,6 +21,7 @@ from pathspec.patterns.gitwildmatch import (
     GitWildMatchPattern,
 )
 import shutil
+import tempfile
 from typing import (
     List,
     Optional,
@@ -42,9 +47,17 @@ def match_file(patterns: List[GitWildMatchPattern], file: str) -> bool:
     return all(matches) if matches else False
 
 
-async def get_namespace(group_name: str, root_nickname: str) -> Optional[str]:
-    path = await pull_namespace_from_s3(group_name, root_nickname)
-    await delete_out_of_scope_files(group_name, root_nickname)
+async def get_namespace(
+    group_name: str,
+    root_nickname: str,
+    presigned_ulr: Optional[str] = None,
+    delete: bool = True,
+) -> Optional[str]:
+    path = await pull_namespace_from_s3(
+        group_name, root_nickname, presigned_ulr
+    )
+    if delete:
+        await delete_out_of_scope_files(group_name, root_nickname)
     return path
 
 
@@ -78,33 +91,67 @@ async def delete_out_of_scope_files(
 
 
 async def pull_namespace_from_s3(
-    group_name: str, root_nickname: str
+    group_name: str,
+    root_nickname: str,
+    presigned_url: Optional[str] = None,
 ) -> Optional[str]:
     local_path = os.path.join(NAMESPACES_FOLDER, group_name, root_nickname)
     os.makedirs(local_path, mode=0o700, exist_ok=True)
 
-    bucket_path = f"s3://continuous-repositories/{group_name}/{root_nickname}"
-    aws_sync_command: List[str] = [
-        "aws",
-        "s3",
-        "sync",
-        "--delete",
-        "--sse",
-        "AES256",
-        "--exact-timestamps",
-        bucket_path,
-        local_path,
-    ]
-    proc = await asyncio.create_subprocess_exec(
-        *aws_sync_command,
-        env={**os.environ.copy()},
-        stdout=asyncio.subprocess.DEVNULL,
-        stderr=asyncio.subprocess.STDOUT,
-    )
-    _, _stderr = await proc.communicate()
-    if _stderr:
-        log_blocking("error", _stderr)
-        return None
+    if presigned_url:
+        async with aiohttp.ClientSession() as session:
+            async with session.get(presigned_url, timeout=None) as response:
+                try:
+                    response.raise_for_status()
+                except ClientResponseError:
+                    log_blocking("error", "failed to download root")
+                    return None
+
+                _, file_path = tempfile.mkstemp()
+
+                with open(file_path, "wb") as file_handler:
+                    while True:
+                        chunk = await response.content.read(8192)
+                        if not chunk:
+                            break
+                        file_handler.write(chunk)
+
+                proc = await asyncio.create_subprocess_exec(
+                    "tar",
+                    "-xf",
+                    file_path,
+                    cwd=os.path.join(NAMESPACES_FOLDER, group_name),
+                )
+                _, _stderr = await proc.communicate()
+                os.remove(file_path)
+                if proc.returncode != 0:
+                    log_blocking(
+                        "error", "failed to decompress repo %s", _stderr
+                    )
+                    return None
+    else:
+        aws_sync_command: List[str] = [
+            "aws",
+            "s3",
+            "sync",
+            "--delete",
+            "--sse",
+            "AES256",
+            "--exact-timestamps",
+            f"s3://continuous-repositories/{group_name}/{root_nickname}",
+            local_path,
+        ]
+        proc = await asyncio.create_subprocess_exec(
+            *aws_sync_command,
+            env={**os.environ.copy()},
+            stdout=asyncio.subprocess.DEVNULL,
+            stderr=asyncio.subprocess.STDOUT,
+        )
+        _, _stderr = await proc.communicate()
+        if _stderr:
+            log_blocking("error", _stderr)
+            return None
+
     try:
         repo = Repo(local_path)
         repo.git.reset("--hard", "HEAD")
@@ -112,4 +159,5 @@ async def pull_namespace_from_s3(
         log_blocking("error", "Expand repositories has failed:")
         log_exception_blocking("exception", exc)
         return None
+
     return local_path
